ETL/etl/pyproject.toml
[tool.poetry]
name = "etl"
version = "0.1.0"
description = ""
authors = ["MATwave <stanislav.matveew@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.7.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
elasticsearch = "^8.7.0"
pydantic = "^1.10.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
